Collection Wheel Finished (?)
Would like to get someone else to verify, but currently the Right Bumper will stat spitting and A with start collection, each one will override one another. To stop, you must press the button you had just pressed. In addition this is all on a toggle.

Must verify with driver team or someone on how they want to control scheme to be (If I remember, we had a hold down to spit one year).