It shoots!
We found out velocity is read at a negative value when we set to a postiive speed.
We tuned our PID since it was way too slow before.
The BackPlate has a programming stop.
At the start of a boot-up, the motors for the X and Y Turret are set to position 0 whever they are