Everything works + Lift Stuff
So I added basic lift code to run two motors in same direction (as Zach said) and to default it to not move.

I fixed the turret we were missing {} and missing a requiremetns for the autoAim Function.

I made it shoot even faster, apparently we can go faster than what I set it to now so.