Auto Started, Shooting Tuning, and Lift
Lift code created and not tested

Adjusted PID Values for Shooter
WE HAVE A BUG IT SPEEDS UP AND THEN SLOWS DOWN
BUt if we re-engage the button while it is at high speeds it keeps up speed and works fine?

SOme code to auto aim Y but we obviously haven't tested

Auto code written and appears on SHUFFLE BOARD but doesn't work lol.