So Practice day and also i forgot to commit code but like it doesn't really matter since i'm the only one uploading (usually)
So like I did this work around that may or maynot work for auto but esssentailly i make commands and put them alone in a parallelracegroup to add the timeout and then add that group into a sequential group to get the timeout for the command.

Also I found out that IsFinished() does actually take in the condition to stop the command

But yeah I got like auto for shooting pathed out but not tested

Also turns out i need to proofread my code and also added position settings for x and y turret.

Slowed the outside intakespeed. We plan to have the trigger to run a quater second more after the limit switch is hit but we'll have to see if my code works